+++ src/app/code/community/EbayEnterprise/Eb2cTax/Test/Model/Order/Create/OrderTest.php
@@ -26,18 +26,20 @@ class EbayEnterprise_Eb2cTax_Test_Model_Order_Create_OrderTest
 	protected $_order;
 	/** @var \eBayEnterprise\RetailOrderManagement\Payload\Order\IShipGroup */
 	protected $_payload;
+	/** @var int */
+	protected $_quoteAddressId = 1;
 
 	public function setUp()
 	{
 		parent::setUp();
 		// replace the session for the logger context
 		$this->_replaceSession('core/session');
-		$this->_taxResponse = $this->getModelMock('eb2ctax/response', ['_construct']);
+		$this->_taxResponse = $this->getModelMock('eb2ctax/response', ['_construct', 'isValid', 'getResponseItemsByQuoteAddressId']);
 		$this->_calculator = Mage::getModel('tax/calculation');
 		$this->_calculator->setTaxResponse($this->_taxResponse);
 		$this->_sdk = Mage::helper('eb2ccore')->getSdkApi('orders', 'create');
 		$this->_payload = $this->_sdk->getRequestBody();
-		$this->_address = Mage::getModel('sales/order_address', ['quote_address_id' => 1]);
+		$this->_address = Mage::getModel('sales/order_address', ['quote_address_id' => $this->_quoteAddressId]);
 		$collection = new Varien_Data_Collection();
 		$collection->addItem($this->_address);
 		$this->_order = $this->getModelMock('sales/order', ['getAddressesCollection']);
@@ -59,7 +61,13 @@ class EbayEnterprise_Eb2cTax_Test_Model_Order_Create_OrderTest
 		$taxQuotes = [Mage::getModel('eb2ctax/response_quote')];
 		EcomDev_Utils_Reflection::setRestrictedPropertyValue($orderItem2, '_taxQuotes', $taxQuotes);
 
-		EcomDev_Utils_Reflection::setRestrictedPropertyValue($this->_taxResponse, '_responseItems', [1 => ['sku1' => $orderItem1, 'sku2' => $orderItem2]]);
+		$this->_taxResponse->expects($this->any())
+			->method('isValid')
+			->will($this->returnValue(true));
+		$this->_taxResponse->expects($this->any())
+			->method('getResponseItemsByQuoteAddressId')
+			->with($this->identicalTo($this->_quoteAddressId))
+			->will($this->returnValue(['sku1' => $orderItem1, 'sku2' => $orderItem2]));
 
 		$handler = Mage::getModel('eb2ctax/order_create_order', ['calculator' => $this->_calculator]);
 		$handler->setTaxHeaderErrorFlag($this->_payload, $this->_order);
@@ -69,15 +77,43 @@ class EbayEnterprise_Eb2cTax_Test_Model_Order_Create_OrderTest
 	/**
 	 * the tax header error flag is true if there are errors
 	 */
-	public function testLoadingTaxDataWithError()
+	public function testLoadingTaxDataWithItemError()
 	{
 		$taxQuotes = [
 			Mage::getModel('eb2ctax/response_quote', ['code' => 'CalculationError']),
 		];
 		$orderItem1 = Mage::getModel('eb2ctax/response_orderitem');
 		EcomDev_Utils_Reflection::setRestrictedPropertyValue($orderItem1, '_taxQuotes', $taxQuotes);
-		// add the response item to the response
-		EcomDev_Utils_Reflection::setRestrictedPropertyValue($this->_taxResponse, '_responseItems', [1 => ['sku1' => $orderItem1]]);
+
+		$this->_taxResponse->expects($this->any())
+			->method('isValid')
+			->will($this->returnValue(true));
+		$this->_taxResponse->expects($this->any())
+			->method('getResponseItemsByQuoteAddressId')
+			->with($this->identicalTo($this->_quoteAddressId))
+			->will($this->returnValue(['sku1' => $orderItem1]));
+
+		$handler = Mage::getModel('eb2ctax/order_create_order', ['calculator' => $this->_calculator]);
+		$handler->setTaxHeaderErrorFlag($this->_payload, $this->_order);
+		$this->assertTrue($this->_payload->getTaxHasErrors());
+	}
+
+	/**
+	 * the tax header error flag is true if there are errors
+	 */
+	public function testLoadingTaxDataWithResponseError()
+	{
+		$taxQuotes = [Mage::getModel('eb2ctax/response_quote')];
+		$orderItem1 = Mage::getModel('eb2ctax/response_orderitem');
+		EcomDev_Utils_Reflection::setRestrictedPropertyValue($orderItem1, '_taxQuotes', $taxQuotes);
+
+		$this->_taxResponse->expects($this->any())
+			->method('isValid')
+			->will($this->returnValue(false));
+		$this->_taxResponse->expects($this->any())
+			->method('getResponseItemsByQuoteAddressId')
+			->with($this->identicalTo($this->_quoteAddressId))
+			->will($this->returnValue(['sku1' => $orderItem1]));
 
 		$handler = Mage::getModel('eb2ctax/order_create_order', ['calculator' => $this->_calculator]);
 		$handler->setTaxHeaderErrorFlag($this->_payload, $this->_order);